requirements.txt
Pillow==9.0.1
numpy
tqdm==4.62.0